[Refactor] 🍺 Repository 패키지 정리
