+++ src/main/java/com/hanaro/schedule_hanaro/admin/service/AdminVisitService.java
@@ -3,8 +3,8 @@ package com.hanaro.schedule_hanaro.admin.service;
 import com.hanaro.schedule_hanaro.admin.dto.response.AdminVisitCarouselDto;
 import com.hanaro.schedule_hanaro.admin.dto.response.AdminVisitNumResponse;
 import com.hanaro.schedule_hanaro.admin.dto.response.AdminVisitStatisticsDto;
-import com.hanaro.schedule_hanaro.customer.repository.CsVisitRepository;
-import com.hanaro.schedule_hanaro.customer.repository.VisitRepository;
+import com.hanaro.schedule_hanaro.global.repository.CsVisitRepository;
+import com.hanaro.schedule_hanaro.global.repository.VisitRepository;
 import com.hanaro.schedule_hanaro.global.domain.CsVisit;
 import com.hanaro.schedule_hanaro.global.domain.Visit;
 import lombok.RequiredArgsConstructor;
