+++ src/main/java/com/hanaro/schedule_hanaro/global/repository/BranchRepository.java
@@ -1,4 +1,4 @@
-package com.hanaro.schedule_hanaro.customer.repository;
+package com.hanaro.schedule_hanaro.global.repository;
 
 import java.util.List;
 import java.util.Optional;
