+++ src/main/java/com/hanaro/schedule_hanaro/global/repository/CsVisitRepository.java
@@ -1,4 +1,4 @@
-package com.hanaro.schedule_hanaro.customer.repository;
+package com.hanaro.schedule_hanaro.global.repository;
 
 import java.time.LocalDate;
 import java.util.List;
