+++ src/main/java/com/hanaro/schedule_hanaro/global/repository/RecommendRepository.java
@@ -1,4 +1,4 @@
-package com.hanaro.schedule_hanaro.customer.repository;
+package com.hanaro.schedule_hanaro.global.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
