+++ src/test/java/com/hanaro/schedule_hanaro/customer/controller/VisitControllerTest.java
@@ -24,10 +24,10 @@ import org.springframework.test.web.servlet.ResultActions;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.hanaro.schedule_hanaro.customer.dto.request.VisitCreateRequest;
-import com.hanaro.schedule_hanaro.customer.repository.BranchRepository;
-import com.hanaro.schedule_hanaro.customer.repository.CsVisitRepository;
-import com.hanaro.schedule_hanaro.customer.repository.CustomerRepository;
-import com.hanaro.schedule_hanaro.customer.repository.VisitRepository;
+import com.hanaro.schedule_hanaro.global.repository.BranchRepository;
+import com.hanaro.schedule_hanaro.global.repository.CsVisitRepository;
+import com.hanaro.schedule_hanaro.global.repository.CustomerRepository;
+import com.hanaro.schedule_hanaro.global.repository.VisitRepository;
 import com.hanaro.schedule_hanaro.customer.service.VisitService;
 import com.hanaro.schedule_hanaro.global.domain.Branch;
 import com.hanaro.schedule_hanaro.global.domain.CsVisit;
